requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
SQLAlchemy==2.0.32
aiosqlite==0.20.0
httpx==0.27.0
python-multipart==0.0.9
requests==2.31.0

fastapi
uvicorn
pydantic
python-dotenv

# --- AI & API Clients ---
requests

# --- Blog/Image Generation ---
pillow
python-docx